Restructure app with multipage support
- Move main app logic to pages/training_backlog.py
- Create new main app.py for multipage navigation
- Add placeholder page2.py for future content
