+++ pages/page2.py
@@ -1,0 +1,13 @@
+import streamlit as st
+
+st.set_page_config(
+    page_title="Page 2",
+    page_icon="📄",
+    layout="wide"
+)
+
+st.title("Page 2")
+
+st.markdown("""
+This is a placeholder page. Replace this content with your own functionality.
+""")
